host: use /proc/self/auxv only when not using cgo
revert partially the previous commit and use /proc/self/auxv only when
cgo is not available.
